+++ cantera/trunk/include/cantera/numerics/GeneralMatrix.h
@@ -163,7 +163,7 @@ public:
      *
      *  @return  Returns a pointer to the top of the column
      */
-    virtual doublereal* ptrColumn(int j) = 0;
+    virtual doublereal* ptrColumn(size_t j) = 0;
 
     //! Index into the (i,j) element
     /*!
@@ -172,7 +172,7 @@ public:
      *
      *  Returns a changeable reference to the matrix entry
      */
-    virtual doublereal& operator()(int i, int j) = 0;
+    virtual doublereal& operator()(size_t i, size_t j) = 0;
 
 
     //! Constant Index into the (i,j) element
@@ -182,7 +182,7 @@ public:
      *
      *  Returns an unchangeable reference to the matrix entry
      */
-    virtual  doublereal operator()(int i, int j) const = 0;
+    virtual doublereal operator()(size_t i, size_t j) const = 0;
 
     //! Copy the data from one array into another without doing any checking
     /*!
@@ -195,7 +195,7 @@ public:
     /*!
      *  We might drop this later
      */
-    virtual  vector_fp::iterator begin() = 0;
+    virtual vector_fp::iterator begin() = 0;
 
     //! Return a const iterator pointing to the first element
     /*!
@@ -211,7 +211,7 @@ public:
      *   @return returns a vector of pointers to the top of the columns
      *           of the matrices.
      */
-    virtual doublereal*   const* colPts() = 0;
+    virtual doublereal* const* colPts() = 0;
 
     //! Check to see if we have any zero rows in the jacobian
     /*!
@@ -222,7 +222,7 @@ public:
      *
      * @return index of the row that is most nearly zero
      */
-    virtual int checkRows(doublereal& valueSmall) const = 0;
+    virtual size_t checkRows(doublereal& valueSmall) const = 0;
 
     //! Check to see if we have any zero columns in the jacobian
     /*!
@@ -233,7 +233,7 @@ public:
      *
      * @return index of the column that is most nearly zero
      */
-    virtual int checkColumns(doublereal& valueSmall) const = 0;
+    virtual size_t checkColumns(doublereal& valueSmall) const = 0;
 
     //! Matrix type
     /*!
